Refactor day 4 with reduce instead of loop
Using the fact that reduce only iterates over all elements. Spares the sentinel nil of the loop.
The function in reduce starts of with an empty list. The empty list is the indicator, that there is an ongoing passport being parsed. All consecutive lines insert their attributes into the current-passport map. Only when there is an empty line, the current-passport is complete, and a new empty list represents the next to-be-parsed passport.
Exploits that
 - (seq "") evaluates to nil.
 - destructuring sets current-passport to nil on an empty line.
 - (into nil {:attr 1}) returns a new seq with [:attr 1] inside